requirements.txt
flask==2.2.3
sqlalchemy==2.0.4
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.0
yfinance==0.2.12
alpha_vantage==2.3.1
mysql-connector-python==8.0.32
python-dateutil==2.8.2
google-generativeai==0.3.1 